Removing  dependencies of fastai for block of datasets.
Also:

- codestyle

- Fix: `to_data` --> `extract_data`.
  `to_data` operate with the instances of `fastai` `ItemBase`
  `extract_data` (renamed version)  operates with  `fastai_sparse` `ItemBase`.
