+++ fastai_sparse/callbacks/__init__.py
@@ -1,3 +1,5 @@
 from .tracker import SaveModelCallback
-from .csv_logger import *
+from .csv_logger import CSVLogger, CSVLoggerIouByCategory
 from .train_val_time import TimeLogger
+
+__all__ = ['SaveModelCallback', 'TimeLogger', 'CSVLogger', 'CSVLoggerIouByCategory']
