+++ fastai_sparse/transforms/__init__.py
@@ -4,7 +4,15 @@ from .main import transfer_keys
 
 from .main import sample_points
 
-from .spatial import *
-from .affine import *
-from .convert import *
-from .colors import *
+from .convert import TfmConvertItem, to_points_cloud, to_sparse_voxels, merge_features
+from .affine import TfmAffine, rotate, rotate_XY, flip_x, scale, rand_scale, translate, rand_translate, noise_affine
+from .colors import TfmColors, colors_noise, colors_normalize
+from .spatial import normalize_spatial, fit_to_sphere, crop_points, elastic
+
+__all__ = ['Transform', 'log_transforms', 'transfer_keys', 'sample_points',
+           'TfmConvertItem', 'to_points_cloud', 'to_sparse_voxels', 'merge_features',
+           'TfmAffine',
+           'rotate', 'rotate_XY', 'flip_x', 'scale', 'rand_scale', 'translate', 'rand_translate', 'noise_affine',
+           'TfmColors', 'colors_noise', 'colors_normalize',
+           'normalize_spatial', 'fit_to_sphere', 'crop_points', 'elastic'
+           ]
