+++ fastai_sparse/transforms/convert.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 from .main import Transform, transfer_keys
-from ..data_items import ItemBase, PointsItem, MeshItem, SparseItem
+from ..data_items import PointsItem, MeshItem, SparseItem
 
 from .. import utils
 
@@ -94,7 +94,8 @@ def _to_sparse_voxels(x: PointsItem):
     points = d['points']
 
     # TODO: is floor better then simply astype(np.int64) ? For x > 0 there is no differences
-    # Some spreadsheet programs calculate the “floor-towards-zero”, in other words floor(-2.5) == -2. NumPy instead uses the definition of floor where floor(-2.5) == -3.
+    # Some spreadsheet programs calculate the “floor-towards-zero”, in other words floor(-2.5) == -2. NumPy instead uses
+    # the definition of floor where floor(-2.5) == -3.
     # >>> a = np.array([-1.7, -1.5, -0.2, 0.2, 0.5, 0.7, 1.3, 1.5, 1.7, 2.0, 2.5, 2.9])
     # >>> b = np.floor(a)
     # >>> c = a.astype(np.int64)
