+++ fastai_sparse/visualize/__init__.py
@@ -1,1 +1,3 @@
 from .ipyvolume import options, scatter, show_mesh
+
+__all__ = ['options', 'scatter', 'show_mesh']
