+++ fastai_sparse/visualize/ipyvolume.py
@@ -16,6 +16,8 @@ from IPython.display import FileLink, display, Image
 from . import utils
 # from ..utils import warn_always
 
+__all__ = ['options', 'scatter', 'show_mesh']
+
 
 @dataclass
 class OptionsOfShow:
@@ -189,7 +191,7 @@ def draw_error_points(points, labels, labels_gt):
     if labels_gt is not None:
         # accuracy
         assert len(labels_gt) == len(labels)
-        accuracy = (labels_gt == labels).sum() / len(labels)
+        # accuracy = (labels_gt == labels).sum() / len(labels)
         # print("Accuracy: {}".format(accuracy) )
 
         # draw errors
@@ -313,12 +315,12 @@ def show_mesh(verts, triangles,
         for label in face_color_dict.keys():
             triangles_set = triangles[face_labels == label]
             color = face_color_dict[label]
-            surface = ipv.plot_trisurf(
+            _ = ipv.plot_trisurf(
                 x, y, z, triangles=triangles_set, color=color)
     else:
         if face_colors is None:
             face_colors = '#f0f0f0'
-        surface = ipv.plot_trisurf(
+        _ = ipv.plot_trisurf(
             x, y, z, triangles=triangles, color=face_colors)
 
     # vertices
