+++ flake8.sh
@@ -1,0 +1,2 @@
+# pip install flake8
+flake8 fastai_sparse/$1
