+++ tox.ini
@@ -1,0 +1,15 @@
+[pycodestyle]
+max-line-length = 160
+ignore = E722,E501
+
+[pep8]
+max-line-length = 160
+ignore = E722
+
+
+[flake8]
+max-line-length = 160
+ignore = E722
+
+
+
